Fix comment about function address replacement
The commit

commit 4313306e0d72236a7c22739590c1182b1d00646c
Author: Gabriel F. T. Gomes <gagomes@suse.de>
Date:   Wed Nov 4 18:57:53 2020 -0300

    Fix call frame information

added a comment explaining how the address of a target function ends up
in register r11. Even though that part of the explanation is correct, it
goes on and incorrectly states that the libraries are modified so that
function calls avoid going through their PLT entries. That part is
utterly wrong (sorry about that), so this patch rephrases the comment.