Fix function interposition
Unlike the libraries in the test suite, running a live-patchable version
of openssl revealed a problem in the interposition of libc functions.
More specifically, openssl has its own constructors, which call some of
the functions that libpulp wants to interpose. Then, if the constructors
in openssl run before those in libpulp, a segmentation fault may occur,
such as demonstrated in the following backtrace:

    Program received signal SIGSEGV, Segmentation fault.
    0x0000000000000000 in ?? ()
    Missing separate debuginfos, use: zypper install openssl-1_1-debuginfo-1.1.1d-0.x86_64
    (gdb) bt
    #0  0x0000000000000000 in ?? ()
 -> #1  0x00007ffff7bd3eb1 in dlopen (filename=0x7ffff7685690 "libcrypto.so.1.1", flags=1) at interpose.c:446
    #2  0x00007ffff75a2514 in get_library_path.constprop () from /usr/lib64/libcrypto.so.1.1
    #3  0x00007ffff75a25ea in verify_checksums () from /usr/lib64/libcrypto.so.1.1
    #4  0x00007ffff75a29cf in FIPS_module_mode_set () from /usr/lib64/libcrypto.so.1.1
    #5  0x00007ffff749a08e in OPENSSL_init_library () from /usr/lib64/libcrypto.so.1.1
    #6  0x00007ffff7de6b3a in call_init.part () from /lib64/ld-linux-x86-64.so.2
    #7  0x00007ffff7de6c46 in _dl_init () from /lib64/ld-linux-x86-64.so.2
    #8  0x00007ffff7dd7f5a in _dl_start_user () from /lib64/ld-linux-x86-64.so.2
    #9  0x0000000000000002 in ?? ()
    #10 0x00007fffffffec01 in ?? ()
    #11 0x00007fffffffec12 in ?? ()
    #12 0x0000000000000000 in ?? ()

Frame 1 shows a call to dlopen, which ends up at address 0, because the
address of the actual dlopen (in libc.so) had not yet been determined by
libpulp. This patch adds checks to the interposing functions.