Livepatch with data layout break
The library entrance tracking feature of Libpulp makes it possible for
live patches to touch multiple functions, and even change the layout of
data shared among them, in a consistent way. This makes live patches
easier to write.

This patch adds a test case that demonstrates that even if a live patch
is crazy enough to change the layout of data shared between two
functions, library entrance tracking guarantees that the live patch
work, provided that both functions (or any other parts of the code that
also use the affected data structures) know about the layout change.