Remove library entrance tracking
Library entrance tracking enables more types of live patches, as
demonstrated by the test suite (contract.py). On the other hand, it
causes problems with algorithms that rely on stack unwinding, such as
thread cancellation and exception handling, which is also demonstrated
by the test suite (cancel.py and exception_handling.py, respectively).

Stack unwinding is affected, because the tracking algorithm steals one
frame from the stack: when a call into a live-patchable library is made,
it doesn't immediately end up in the target function; instead, it goes
through __ulp_entry, which saves the return address into TLS storage,
then replace it on the stack with an address of its own. Since the
original return address is no longer on the stack, unwinding cannot go
any further, which causes the problems reported in the commits
12ea589ab9c3 and d645ee8c5e5d (see their commit messages for detailed
information).

This patch disables library entrance tracking altogether, which solves
the unwinding problem, but, as side-effect, makes it harder to create
live patches (the burden of creating sane live patches will then lays on
the people writing them).

Notably, the removal of library entrance tracking makes it easier to
make libraries live patchable. Previously, their build systems should be
changed to: 1. include trm.S/trm.o in the linking phase; 2. expose
Libpulp symbols globally; 3. add -fpatchable-function-entry to the
compiling phase; now, only (3) remains. Unfortunately, the presence of
symbols from Libpulp was used to determine whether a library was live
patchable or not, thus, after this patch, live patchable libraries are
no longer detect by the ulp tool.

Finally, after this patch, the test cases for thread cancellation and
exception handling pass, whereas the blocked.py and contract.py test
cases start to fail, precisely because they were designed to detect
problems with the migration to new universes on a per-thread basis.