Improve error path
This patch adds more options to the main tools used in live patch
application: the Trigger and the Check tools. The '-v' options make
these tools print useful information about the live patching process.
Conversely, the '-q' option hides most warning messages, which might be
useful for scripted use of the tools. However, even when this option is
in effect, not all messages are omitted; fatal errors still produce some
output, to let the user know that the target process might have been put
into an inconsistent state, and should probably be terminated.