Fix Check tool
The Check tool receives a metadata filename as its second command-line
argument; then parses the live patch id out of it; then writes this id
to the target process memory space; then, finally, calls libpulp
routines to determine whether the referred live patch has already been
applied to the target process or not.

However, __ulp_check_applied_patch (a function from libpulp.so)
mistakenly used __ulp_metadata_ref (a global variable) uninitialized,
leading to a null pointer dereference in:

  lib/ulp.c line 1096:
    if (memcmp(patch->patch_id, id, 32) == 0)

The fault was caught be the tracer program and a bogus result was
returned.