Packer tool uplift
Use argp in the packer tool so that it provides more options, such as
the ability to be extra verbose to help with debugging, or completely
silent if requested. Also make the error path print more warnings about
what went wrong. Using argp also gives a more standard usage message.