Do not use randomly generated numbers as live patch ID
This patch removes the use of a randomly generated number as the live
patch ID. Now, the live patch is composed of two parts: the first half
is taken from the build ID of the live patch DSO; the second half is a
MD4 hash of the description file contents. This makes live patch IDs
deterministic, which avoids reproducibility concerns.

The MD4 implementation is based on the description provided by IETF in
RFC 1320 [1], however, it does not use the reference implementation. As
a matter of fact, it even yields different results, because the bit
rotation is implemented on 32-bits variables, rather than on unsigned
long int variables, which may have 64 bits on some architectures.

[1] https://tools.ietf.org/html/rfc1320