Add test case to show that exception handling is also broken
Similar to the commit

commit 12ea589ab9c3ba492c68588280aac5d1958db580
Author: Gabriel F. T. Gomes <gagomes@suse.de>
Date:   Thu Mar 18 17:57:30 2021 -0300

    Add test case for thread cancellation problem

this patch adds a new test case to show that exception handling is also
affected by the frame stealing in __ulp_entry. The test program is
extremely simple: it calls into a c++ library that throws a generic
exception. The exception is first caught within the library, then
rethrown, and finally caught in the main program.

The test case has no problem handling the exception within the library,
and the first 'catch' block (in the 'outer' function) works correctly.
However, after the exception is rethrown, the stolen frame messes with
the unwinding of the stack, and the second 'catch' block (in 'main')
never gets triggered.

Attaching to the target process reveals that the unwinding mechanism
loops forever (the top items of the backtrace change every time the
process is attached to; this is just one example):

  #0 in execute_cfa_program () at libgcc/unwind-dw2.c:1154
  #1 in uw_frame_state_for () at libgcc/unwind-dw2.c:1311
  #2 in _Unwind_RaiseException () at libgcc/unwind.inc:104
  #3 in __cxxabiv1::__cxa_throw (<std::runtime_error::~runtime_error()>) at libstdc++-v3/libsupc++/eh_throw.cc:90
  #4 in outer () at tests/libexception.cc:18
  #5 in __ulp_entry () at lib/trm.S:184
  Backtrace stopped: frame did not save the PC

PS: the commit message of commit ID 12ea589ab9c3 incorrectly states that
the affected thread is hanging in the unwinder, when, in fact, it is
also looping forever. Letting the thread continue for a while, then
inspecting again, would reveal that the unwinder was in a different
place, thus busy working on the broken data.