Check the stack before applying a live patch
As exposed by the test suite, stealing a frame from the stack to track
library entrance breaks thread cancellation and exception handling.
Therefore, this patch adds an alternative solution, which uses libunwind
to inspect the call stack of all threads in the target process, before
actually applying the live patch. If any of the frames of any of the
threads has the program counter within the memory space of the target
library, the live patching operation is aborted. On the other hand, if
no frames lie within the library, live patching can proceed with the
same consistency properties that the library entrance tracking algorithm
provided.