Fix compilation errors with clang
If requested, Libpulp can now be built with clang. Most of the changes
are cosmetic, though, and clang did not reveal any new warnings or
errors. The -fno-inline option in the test suite avoids inlining of
intra-library calls (without it, clang automatically inlines calls in
libblocked and libcontract, leading to unexpected passes (XPASS) in the
test suite - gcc does not inline the calls, but the new flag is useful
for gcc, too; in the sense that it might avoid future problems if gcc
changes its default settings).