New test case with missing function in the live patch
A live patch is mainly composed of a metadata file and a shared library.
The library contains new functions to replace older, buggy functions on
the target process; whereas the metadata file contains the names of
replaced and replacing functions, in pairs. When a live patch is
applied, Libpulp loads the shared library, then uses the names contained
in the metadata to find the in-memory addresses of the affected
functions.

However, the metadata file could be wrong and reference functions that
do not exist in the live patch, perhaps because of a typo. When that
happens, Libpulp fails with error messages similar to the following:

  ulp: Unable to load function foo: livepatch.so: undefined symbol: foo.

This error is probably not critical and the target process could resume
execution. However, it leaves the shared library loaded in the target
process memory, polluting its namespace. This can be observed in the
/proc filesystem:

  $ cat /proc/PID/maps
  ...
  7f1000-7f2000 r--p 00000000 00:2c 4191 livepatch.so.0.0.0
  7f2000-7f3000 r-xp 00001000 00:2c 4191 livepatch.so.0.0.0
  7f3000-7f4000 r--p 00002000 00:2c 4191 livepatch.so.0.0.0
  7f4000-7f5000 r--p 00002000 00:2c 4191 livepatch.so.0.0.0
  7f5000-7f6000 rw-p 00003000 00:2c 4191 livepatch.so.0.0.0
  ...

Since the namespace pollution is undesirable, this patch adds a new test
case, which demonstrates the problem. The test case is marked as
expected to fail until it gets fixed.