Only use function version selection routine in live patching time
Until the commit

commit f85145013ca2f10b2e001dcea56d5eab0143d9bd
Author: Gabriel F. T. Gomes <gagomes@suse.de>
Date:   Thu Apr 1 16:19:29 2021 -0300

    Remove library entrance tracking

each thread migrated to new versions of live patched functions on their
own, after each of them crossed the application-library boundary. Thus,
selecting between function versions had to be done during the regular
execution of the application. In order for this mechanism to work, a
live patched function would have its prologue modified, so that the
execution got diverted to a function version selection routine, which
would use a thread-local universe counter to perform the selection. The
following snippet shows the code layout before and after live patching:

  Before live patching:

    function prologue:
        nop
        nop
        ...
    function entry:
        nop
        <function body>

  After live patching:

    function prologue:
  ┌───> push    %rdi
  │     movq    $index, %rdi
  │     jmp     0x0(%rip)
  │     <data>  <function version selection routine>
  │ function entry:
  └──── jmp     <function prologue>

However, since the commit above, all threads migrate to new versions of
live patched functions at the same time, during live patch application.
Thus, there is no need to go through the selection routine anymore,
because the address of the target function can be hardcoded into the
<data> area of the prologue. After this patch, the prologue looks like
the following after a patch has been applied:

    function prologue:
        push    %rdi           <-- Unused!
        movq    $index, %rdi   <-- Unused!
  ┌───> jmp     0x0(%rip)
  │     <data>  <target function>
  │ function entry:
  └──── jmp     <function prologue>

Notice how the first two instructions are no longer used, because their
purpose, i.e. to set the argument to the function version selection
routine, is no longer required.

The main benefit of hardcoding the target address of the live patched
function in the prologue is the fact that, during the regular execution
of the application, the function version selection routine is not
executed, which improves performance, specially when the affected
library function is short and called often.

The change described above is sufficient to make live patching work
without using the function version selection routine during the regular
execution of the application. However, it breaks live patch reversal. To
fix that, this patch adds more logic to libpulp, so that, whenever a
patch gets reverted, the <data> area of the prologue gets updated, and
that happens during live patch application.

Notes about the unused instructions in the prologue:

Since the first two instructions in the prologue of live patched
functions are no longer used, the prologue itself may shrink, so may
argument to -fpatchable-function-entries. This patch does not implement
that, but a future patch should.