Ensure bootstrap check if user has requirements
The `boostrap` only runs the necessary commands in order to generate
the `configure` script, but do not check if the user indeed have the
necessary files installed. This commit fixes that by checking if
every command can be run before executing the bootstrap process, and
on failure display a single message showing which command has failed.

Signed-off-by: Giuliano Belinassi <gbelinassi@suse.com>