Avoid opening libdl.so to interpose dlsym
Previously, libpulp opened and parsed the ELF content of libdl.so file
to find the address to dlsym to interpose it. Instead of relying on the
.so file, we parsed the alreay loaded ELF program header structures that
is in the program memory to find the address to dlsym.

This commit also removes the requirement of libelf from libpulp.so, but
the tools still require it.

Signed-off-by: Giuliano Belinassi <gbelinassi@suse.de>