Add revert-all for trigger
A new option is available for trigger: --revert-all. Usage:

$ ulp trigger -p $pid --revert-all $libname $metadata

--revert-all causes trigger to revert all patches applied to $libname in
the target process with pid $pid.

In case --revert-all is provided, $metadata is optional. In case the
metadata is provided with patch content, the new patch will be atomically
applied to the process after all older patches have been reverted, so
the target process doesn't get a chance to run the unpatched version of
the function. If $metadata is not provided, no patch will be applied
after removal.

Signed-off-by: Giuliano Belinassi <gbelinassi@suse.de>
Co-authored-by: Simon Lees <sflees@suse.de>