Add initial support to doxygen
Doxygen is a Javadoc-like document generator compatible with several
languages, including C. It works by reading the comments in function
headers, which must follow certain patterns, and automatically
generates webpages and TeX files with documentation.

Currently, some functions in libpulp are almost doxygen compatible,
but others need some work explaining what they do. In the end,
this documentation may be used to newcommers to have a more holistic
view to the project, and on release, provide an documentation to
Libpulp's ABI for end users.

Signed-off-by: Giuliano Belinassi <gbelinassi@suse.de>