Add support for relative paths in target library
Now the metadata format accepts relative paths when specifying the path
to the target library. The way it works is by comparing both library
base name and build id to match the library to patch, instead of relying
on the full path to it.

Example:
```
__ABS_BUILDDIR__/.libs/libparameters_livepatch1.so
@.libs/libparameters.so.0
int_params:new_int_params
float_params:new_float_params
```
Notice that the line starting with @ do not require to be the full path
to target library anymore.

Signed-off-by: Giuliano Belinassi <giulianob@gbelinassi.udp.ovpn1.nue.suse.de>