Add ulp_apply tool for packaging
This is a script that we will call from rpm post package install
to update any running processes with the new live patches.

While it could still use things such as better error reporting at
this point it is functional. The name and the final location for
the script are open to change (currently its not installed)