Block livepatching if fatal error is found
If a fatal error was found when livepatching, we disable the
livepatching mechanism instead of exiting the target process.

The only case when we crash the target process is if libpulp
initialization fails. It may be better to warn the user that the
livepatching capabilities is not working on process startup rather than
silently disabling it and then some sysadmin finds out that something
went wrong weeks later.

Signed-off-by: Giuliano Belinassi <gbelinassi@suse.de>