Fix (some) race conditions in testsuite
The testsuite had some race conditions due to the new manyprocesses.py
test. There is also timeout issues which failed depending of how much
CPU is being used at the testing moment. In this commit we improve that
by:

1- Serializing manyprocesses.py, which now is run when all other tests
   are complete.
2- Increase redzone.py timeout to 50, avoiding failure when the CPU is
   too busy.

Other problems may still be around, but so far I have not found any.

Signed-off-by: Giuliano Belinassi <gbelinassi@suse.de>