Fix hidden.py test failure in clang
hidden.py were set to XFAIL because it required access to symbols not
available in .dynsym. However, we implemented access to it because it
was required for standalone process livepatching (i.e., patching
functions outside libraries), thus this test now is accepted.

However, clang removes the hidden_dozen symbol completely from the
program because inline that function is always a good idea for both
performance and code size reasons, and it removes the symbol completely
from the .symtab as well. Adding the volatile forces clang to at least
generate a jmp to hidden_dozen and the symbol is not removed from
.symtab.

Signed-off-by: Giuliano Belinassi <gbelinassi@suse.de>