Fix deadlock testcase
Deadlock testcase could fail depending of CPU usage because it can't
trigger the patch when the program is unlocked. There is no way to solve
this issue generally (i.e. it is equivalent to the halting problem), so
only fails if trigger indeed deadlocks and never returns, exposing a
deadlock in libpulp. If all attempts failed it should still be a
success because it isn't a deadlocks.

Signed-off-by: Giuliano Belinassi <gbelinassi@suse.de>