Fix self-querrying
If the `ulp` tool is loaded with libpulp.so, livepaches capabilities are
loaded with it and the process turns to be livepatchable. However,
ptrace cannot be used to attach to itself, so the livepatching mechanism
does not work. So avoid failure by simply not adding itself to the
process list.

Signed-off-by: Giuliano Belinassi <gbelinassi@suse.de>