Integrate metadata into ulp container
This commit drastically changes how the patch files are distributed.
Instead of having three files (an .ulp, a .rev and a .so), we now have
only one (a .so). The content of every other file is integrated into .so
file through the sections .ulp and .ulp.rev. These sections are
extracted on trigger time.

This way, instead of passing a .ulp or .rev to trigger we now only have
to pass the .so to it.

Trigger also changes. Now, to revert the patch, the user has to pass
--revert flag.

Signed-off-by: Giuliano Belinassi <gbelinassi@suse.de>