Improve trigger messages
Trigger messages can be quite complicated to understand because of the
huge amount of information displayed on screen.

We now refactor this to provide more simplified messages:

For instance, when triggering all processes with the patches in current
directory:

```
$ ulp trigger '*.so'

Summary:
  ulp-test (pid: 25606):
    SKIPPED ./cve-2022-0778_livepatch1-1.1.1l-150400.2.47.so: Build ID mismatch
    SKIPPED ./cve-2022-0778_livepatch1-1.1.1l-150400.3.4.so: Build ID mismatch
    SKIPPED ./cve-2022-0778_livepatch1-1.1.1l-150400.3.11.so: Build ID mismatch
    SKIPPED ./cve-2022-0778_livepatch1-1.1.1l-150400.3.3.so: Build ID mismatch
    SKIPPED ./cve-2022-0778_livepatch1-1.1.1l-150400.3.15.so: Build ID mismatch
```

Or when asking to revert all patches to libcrypto.so to all processes:
```
ulp trigger --revert-all=libcrypto.so

Summary:
  ulp-test (pid: 25606):
    SUCCESS reverted all patches from libcrypto.so
```

Or when a trigger to a single process fails with build id mismatch:

```
$ ulp trigger -p $(pidof ulp-test) cve-2022-0778_livepatch1-1.1.1l-150400.3.15.so

error: could not apply cve-2022-0778_livepatch1-1.1.1l-150400.3.15.so to ulp-test (pid 25606): Build ID mismatch
note: run `ulp patches -b` to retrieve all build ids from patchable processes.
```

This should improve readability all across the board.

Signed-off-by: Giuliano Belinassi <gbelinassi@suse.de>