Add support to functions starting with endbr64 instruction
Some functions can begin with `endbr64` in order to validate indirect
jumps in x86_64, even if -fpatchable-functions-entry is enabled on
compilation time.

For instance, on Tumbleweed's glibc, it can have:

__aio_read:
  endbr64
  nop
  nop
  ...

As you can see, the endbr64 insn is placed before the nop opcodes.
Previosly, trying to patch this function would result in bad insn,
as it expected the two nop's to be the two first insn of __aio_read.
This patch makes libpulp check if endbr64 is present there and write
the redirect after it.

Signed-off-by: Giuliano Belinassi <gbelinassi@suse.de>