Do not interpose dlsym
Glibc's dlsym behaviour changes depending of where it is being called.

Proof of this can be found in glibc's dlfcn/dlsym.c. Notice that
RETURN_ADDRESS(0) is passed to internal functions, which gets the
previous frame address.

The practical behaviour is that if we interpose dlsym, then software
that interposes functions from glibc will find libpulp functions
instead, and libpulp will find the software interposed functions,
creating an infinite recursion.

Therefore, we can not interpose dlsym and use a custom lock, as we do
for the other functions, and we instead check if the dlsym lock is
held.

Signed-off-by: Giuliano Belinassi <gbelinassi@suse.de>