Do not rely on mmap when asunsafe_begin wasn't called
The previous implementation had the following problem:

Suppose that a global constructor that runs before aunsafe_begin. This
means that real_malloc is null and it would fall back to mmap. Now,
after asunsafe_begin is run and real_malloc is known, calling free would
call real_free, resulting in a crash because the memory chunk was allocated
with mmap.

Therefore, instead of using mmap, we link against __libc symbols to
escape the interpose problem.

Signed-off-by: Giuliano Belinassi <gbelinassi@suse.de>