Add support for AFL
AFL -- American Fuzzer Loop -- is an input fuzzer which takes valid
intputs for a program and mutates them in order to find bugs.

AFL can be enabled in this project by using

configure --enable-afl-check
cd tests
make check-afl

Those tests can run for ages, so it is not advised to run them to check
if a particular feature is bug-free, but rather use this tool in a short
period of time to find silly input related bugs.

New tests can be added on tests/fuz/

Signed-off-by: Giuliano Belinassi <gbelinassi@suse.de>