Add timeout to ptrace
The `ulp` tool uses ptrace to both update a process AND query for libraries
and patches installed in the process, and both things could be happening at
once. If this is the case, ptrace may sometimes fail with permissions or
busy errors. In this case, this wrapper tries ot ptrace a process multiple
times before concluding that it can't.

Signed-off-by: Giuliano Belinassi <gbelinassi@suse.de>