Refactor setup_package.sh script
A rewrite of the setup_package.sh script. This version uses the
information available in download.suse.de and crawl there to get the
necessary packages: the binaries, the source file, and the ipa-clones
artifact.

Crawling has the advantage of being clear for someone maintaining the
script what should be parsed and get, rather than relying on
undocumented or obscure osc api features.