Add -R --root=PREFIX option
When `ulp` tool runs in a chroot'ed directory, the `realpath` function
will return the path to the root from the chroot, not the path from
the true root.  This commit introduces the --root=PREFIX options which
appends PREFIX to the path to the patch file right before sending it
to the target process, providing a way to workaround this chroot
problem.

Signed-off-by: Giuliano Belinassi <gbelinassi@suse.de>