Add support for target library dumps
In order to avoid requiring the original library to build the
livepatches, we now support the use of dumps of the target library as
input. This should remove the need to package the original libraries
into tarballs when building the livepatches.

The process can now be done by:

1- Running:
```
$ ulp extract targetlib.so -o targetlib.so.json
```

2- Specify the .json file as a livepatch target in the dsc file:

```
 liblivepatchcontainer_livepatch1.so
 @path/to/liblivepatch.so.json
 ...
```

The old way of specifying the library itself also works and is
provided as compatibility.

Signed-off-by: Giuliano Belinassi <gbelinassi@suse.de>