Fix linkage of libunwind and file descriptor leak
Previously, libpulp.so.0 were being linked against libunwind, but no
function of this library is called inside libpulp. This library also
do not have any constructors. Hence, only the tools need to link
against this library.

Also fix a file descriptor leak in the stack check code. Since it is
not used in the package updater, this is not critical.

Signed-off-by: Giuliano Belinassi <gbelinassi@suse.de>