Fix `patches` not detecting libcrypto as livepatchable
libcrypto.so.1.1 required to read around 200 symbols in order for
`patch` to detect it as a livepatchable library.  Hence we bump
the number of symbols to read from 64 to 8000.  This value of 8000 is
found by analyzing all libraries found in /usr/lib64 and its subfolders
for how many symbols it would need to be read in order to decide if
the library is livepatchable.

Workarround #159

Signed-off-by: Giuliano Belinassi <gbelinassi@suse.de>