Add mechanism to disable livepatching on processes
The following new command:

ulp set_patchable -p PROCESS <enable/disable>

to enable or disable livepatching of a target process. PROCESS can be
a wildcard matching a set of processes, or a number describing a PID.

Example:

$ ulp set_livepatchable -p numserv disable

output:

ulp: Process numserv (pid: 5653): livepatching is now disabled.

This can be used in the future to enable livepatching on a process that
was previously set to disabled.

Signed-off-by: Giuliano Belinassi <gbelinassi@suse.de>