Detect if mprotect is blocked when libpulp is initalized
Processes launched by systemd can specify:

MemoryDenyWriteExecute=yes

to block redefining page flags with EXEC flag. This is a problem for
livepatching, once we need to redefine the page flags in order to
install the redirection jump.

This commit adds a mechanism to detect this case and disable livepatch
accordingly, as we don't have the necessary engine features yet to
circunvent this problem from `ulp` side.

Signed-off-by: Giuliano Belinassi <gbelinassi@suse.de>