Fix livepatch status counting
Previously, the output of `ulp trigger` counted the number of prcoesses
that were patched skipped and failed, but this counting did break at
times because it only remembered the last output of trigger_one_process.
Now it logs into a bitfield to log if there was a success skip or
failure.

Fixes #177

Signed-off-by: Giuliano Belinassi <gbelinassi@suse.de>