Replace dlsym with ELF file parsing for checking for symbol existence
Previously, a sanity check in the `ulp` tool checking if the symbols
declared in the description section is actually in the code section.
This check was done by loading the patch with dlopen. However some
security flags in openbuildservice introduces a different behaviour
on dlopen, turning it unreliable to check this.

Now we parse the file as a ELF and check if the symbol is in the
symbol table.

Signed-off-by: Giuliano Belinassi <gbelinassi@suse.de>