Change .so load policy and fix symbol visibility problems
Previously, libpulp used RTLD_LAZY in dlopen, which resulted in
livepatch that should have failed to apply being applied. In SLE-15-SP4,
those patches seems to work regardless of that, but in SP5 it result in
problems. Hence, change this policy to RTLD_NOW to force all symbols
to be loaded by the linker at patch load time.

Also, this commit now encode the symbol references on packer even if
not specified in .dsc by capturing it from the .json dumps, which
should improve the support for strong externalizations.

Signed-off-by: Giuliano Belinassi <gbelinassi@suse.de>