Use new -fpatchable-function-entry layout
PowerPC -fpatchable-function-entry now generates the correct layout.
Update libpulp to reflect this.

Now the redirection works more like how it is supposed:
 - Generate 18 nops before the function entry point to contain the trampoline
 - Generate 1 nop after the function entry point to contain the branch to the trampoline

This was we don't need any address hashing or remembering which
instructions were there to remove the livepatch. However, since the
compiler can remove the global entry point of the function, we still
have to do some checking to see exactly where the NOP is.

Signed-off-by: gbelinassi <gbelinassi@suse.de>